+++ src/main/java/subway/ui/line/LineController.java
@@ -1,4 +1,4 @@
-package subway.controller.line;
+package subway.ui.line;
 
 import java.net.URI;
 import java.util.List;
@@ -10,10 +10,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import subway.controller.dto.AddLineRequest;
-import subway.controller.dto.AddLineResponse;
-import subway.controller.dto.LineResponse;
-import subway.service.line.LineService;
+import subway.application.line.LineService;
+import subway.ui.dto.AddLineRequest;
+import subway.ui.dto.AddLineResponse;
+import subway.ui.dto.LineResponse;
 
 @RequiredArgsConstructor
 @RestController
@@ -27,13 +27,13 @@ public class LineController {
         final AddLineResponse addLineResponse = lineService.addLine(request);
         final URI uri = URI.create("/lines/" + addLineResponse.getId());
         return ResponseEntity.created(uri)
-                .body(addLineResponse);
+            .body(addLineResponse);
     }
 
     @GetMapping
     public ResponseEntity<List<LineResponse>> getLines() {
         final List<LineResponse> lineResponses = lineService.getLines();
         return ResponseEntity.ok()
-                .body(lineResponses);
+            .body(lineResponses);
     }
 }
