+++ src/main/java/subway/ui/station/StationController.java
@@ -1,4 +1,4 @@
-package subway.controller.station;
+package subway.ui.station;
 
 import java.net.URI;
 import javax.validation.Valid;
@@ -9,10 +9,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import subway.controller.dto.AddStationRequest;
-import subway.controller.dto.AddStationResponse;
-import subway.controller.dto.RemoveStationRequest;
-import subway.service.line.LineService;
+import subway.application.line.LineService;
+import subway.ui.dto.AddStationRequest;
+import subway.ui.dto.AddStationResponse;
+import subway.ui.dto.RemoveStationRequest;
 
 
 @RequiredArgsConstructor
@@ -27,13 +27,13 @@ public class StationController {
         final AddStationResponse addStationResponse = lineService.addStation(request);
         final URI uri = URI.create("/stations/" + addStationResponse.getStationId());
         return ResponseEntity.created(uri)
-                .body(addStationResponse);
+            .body(addStationResponse);
     }
 
     @DeleteMapping
     public ResponseEntity<Void> deleteStation(@RequestBody @Valid final RemoveStationRequest request) {
         lineService.removeStation(request);
         return ResponseEntity.noContent()
-                .build();
+            .build();
     }
 }
